Made Track control and effect insert components which can simply be passed the player and recorder to operate. Setting up to implement context.
